refactor(server): consolidate and relocate custom error classes
