+++ packages/server/src/skills/errors.ts
@@ -1,25 +1,32 @@
-export class SkillNotFoundError extends Error {
+class SkillError extends Error {
+  constructor(message: string) {
+    super(message);
+    this.name = 'SkillError';
+  }
+}
+
+export class SkillNotFoundError extends SkillError {
   constructor(name: string) {
     super(`Skill "${name}" not found`);
     this.name = 'SkillNotFoundError';
   }
 }
 
-export class SkillNameCollisionError extends Error {
+export class SkillNameCollisionError extends SkillError {
   constructor(name: string) {
     super(`Skill name "${name}" already exists`);
     this.name = 'SkillNameCollisionError';
   }
 }
 
-export class SkillInvalidError extends Error {
+export class SkillInvalidError extends SkillError {
   constructor(message: string) {
     super(message);
     this.name = 'SkillInvalidError';
   }
 }
 
-export class SkillImportError extends Error {
+export class SkillImportError extends SkillError {
   constructor(message: string) {
     super(message);
     this.name = 'SkillImportError';
