+++ packages/server/src/tools/errors.ts
@@ -113,46 +113,3 @@ export class WebFetchResponseTooLargeError extends ToolError {
     this.limitBytes = limitBytes;
   }
 }
-
-export class BrowserBridgeNotConnectedError extends ToolError {
-  constructor() {
-    super('Desktop browser bridge is not connected. Browser tools require the Stitch desktop app.');
-    this.name = 'BrowserBridgeNotConnectedError';
-  }
-}
-
-export class BrowserBridgeNotConfiguredError extends ToolError {
-  constructor() {
-    super('Browser tools require the Stitch desktop app. No desktop browser bridge is configured.');
-    this.name = 'BrowserBridgeNotConfiguredError';
-  }
-}
-
-export class BrowserSessionNotSetError extends ToolError {
-  constructor() {
-    super('Browser sessionId must be set before executing commands.');
-    this.name = 'BrowserSessionNotSetError';
-  }
-}
-
-export class BrowserMissingFieldError extends ToolError {
-  readonly field: string;
-  readonly tool: string;
-  constructor(tool: string, field: string) {
-    super(`Missing required field: ${field} for tool ${tool}`);
-    this.name = 'BrowserMissingFieldError';
-    this.field = field;
-    this.tool = tool;
-  }
-}
-
-export class BrowserInvalidOpError extends ToolError {
-  readonly op: string;
-  readonly tool: string;
-  constructor(tool: string, op: string) {
-    super(`Invalid op for ${tool} tool: ${op}`);
-    this.name = 'BrowserInvalidOpError';
-    this.op = op;
-    this.tool = tool;
-  }
-}
